Clean and reformat - Moved file predicates to FileChooser - As a result, remove unused/empty StorageLocation
